I don't wanna talk about it 🙃
So. I had to do a complete restructure of the entire project. We now have the player as a singleton autoload, there's two scenes where the player cannot be seen. This is the main menu and the galaxy map. So, after rigging the main menu (AGAIN!) to handle the new logic, I moved on to the Galaxy Map.

In the Galaxy Map I hooked the pause button up to the datakey_collected function in the Protoship and gave myself +=1 every time. I then transfered this data to the Galaxy Map and wired it up to a pathnode2d -> follownode2d and then translated the values of the follownode2d's progress_ratio to datakey milestones. So every time the player collects a datakey, a sprite that looks like the protoship is snapped to the correct progress_ratio. So yeah, we literally have player progression, and what?

Next, I went to the datakey and restructured their logic to ignore the scene_manager script I had made yesterday. I then calibrated the GameManager to work with the Protoship autoload and not a prefab scene we throw in there and bingo, it works like a charm. I hooked each button up to the test_hell for now.

I also introduced a secondary spawner that has its own frequency that doesn't start ticking until the player passes an initial threshold (minimum 30 seconds please ffs 🤣) and then this all works with the boss triggers I set up.

FINALLY!

I fixed the massive gapping bug that's plagued my life since day two. Player's will now take damage on the FIRST SHOT!!! YES YOU HEARD ME! no more cheesing the first shot!

This mean I can FINALLY spend a WHOLE day on enemies and enemies alone.

I'm actually looking forward to it a lot!

~Tat